% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.data.sim.R
\name{cn.simulate.data}
\alias{cn.simulate.data}
\title{Simulating copy number for a collection of trios.}
\usage{
cn.simulate.data(n, p, theta, sigma)
}
\arguments{
\item{n}{The number of trios.}

\item{p}{A vector representing the probabilities of each copy number state.}

\item{theta}{A vector representing the mean LRR of each copy number state.}

\item{sigma}{A vector representing the average LRR variance for each copy number state.}
}
\value{
A collection of trios with their copy number genotypes.
}
\description{
This function allows you to simulate copy number for a dataset of trios. This can be used as input for the gibbsSamplerMendelian function.
}
\examples{
data.1 <- simulate_data(66, p = p.1000gp, theta = theta.1000gp, sigma = sigma.1000gp)
saveRDS(data.1, "simulatedData.rds")
}
\keyword{marimba}
