% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_cn2}
\alias{map_cn2}
\title{Return maximum a posterior copy number from a model}
\usage{
map_cn2(model)
}
\arguments{
\item{model}{a fitted model}
}
\value{
a vector of the maximum a posterior copy number
}
\description{
Return maximum a posterior copy number from a model
}
