% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data from simple deletion model - Genetic model}
\usage{
simulate_data(params, N, error = 0)
}
\arguments{
\item{N}{number of trios to simulate}

\item{error}{probability of non-Mendelian event in offspring}

\item{region}{integer indexing the 1000G data}
}
\description{
Simulate data from simple deletion model - Genetic model
}
